specify content type for metadata.
Entities MAY publish their metadata documents at a well known location by placing the document at the
location denoted by its unique identifier, which MUST be in the form of a URL (rather than a URN). See
Section 8.3.6 of [SAMLCore] for more information about such identifiers. It is STRONGLY
RECOMMENDED that https URLs be used for this purpose. An indirection mechanism supported by the
URL scheme (such as an HTTP 1.1 302 redirect) MAY be used if the document is not placed directly at
the location. If the publishing protocol permits MIME-based identification of content types, the content type
of the metadata instance MUST be application/samlmetadata+xml.

https://docs.oasis-open.org/security/saml/v2.0/saml-metadata-2.0-os.pdf